isssue #12 - Fix null handling - Through testing of the RecruiterController, found we were NOT properly handing errors when accessing the database as it will NOT happen in live - This has been fixed to allow testing to pass on the Recruiter Controller methods

+++ RecruiterPathway/Controllers/RecruitersController.cs
@@ -49,6 +49,10 @@ namespace RecruiterPathway.Controllers
         {
             //Find the matching user from the DB
             var result = await repository.SignInRecruiter(model);
+            if (result == null)
+            {
+                return StatusCode(500);
+            }
             //Check if user exists and if password is valid
             if (result)
             {
@@ -261,8 +265,20 @@ namespace RecruiterPathway.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
-            var self = await repository.GetSignedInRecruiter(HttpContext.User);
-            var recruiter = await repository.GetById(id);
+            Recruiter self;
+            if (HttpContext != null)
+            {
+                self = await repository.GetSignedInRecruiter(HttpContext.User);
+            }
+            else
+            {
+                self = await repository.GetById(id);
+            }
+            if (self == null)
+            {
+                //Something didn't work right with the database, return a server error
+                return StatusCode(500);
+            }
             if (id == self.Id || self.Email == "administrator@recruiterpathway.com")
             {
                 repository.Delete(id);
